Added cache busting.
js files used a versioned resource strategy, and all other files will
use a MD5 hash to determine if the file has changed. A hard-coded
boolean controls if the cache is enabled or not. This could easily be
changed to be from a properties file instead.

I added in jquery and bootstrap just to have some things to pull in. The
glyphicon used shows how Spring auto-converts internal css links to use
the MD5 hash as well.